Update Arithmetic progression.py

+++ Arithmetic progression.py
@@ -4,11 +4,11 @@ import math
 print('Arithmetic progression')
 print('use to find Sn')
 print('Sn =(n/2*(2*a1+(n-1)*d)')
-a1 = int(input('a1(the initial term) : '))
+a1 = float(input('a1(the initial term) : '))
 n = int(input('n : '))
-d = int(input('d(the common difference) : '))
+d = float(input('d(the common difference) : '))
 
 def Sequences(n):
-    return int(n/2*(2*a1+(n-1)*d))
+    return float(n/2*(2*a1+(n-1)*d))
 
 print('Sn = ' + str(Sequences(n)))
